refactor: color, basic 질문지 분리
