+++ src/pages/question.tsx
@@ -1,22 +1,14 @@
 import Image from 'next/image';
 import { useRouter } from 'next/router';
-import type { ChangeEvent } from 'react';
 import { useState } from 'react';
 import styled from 'styled-components';
 
-import GradientBox from '@/component/common/GradientBox';
 import withLayout from '@/component/hoc/withLayout';
-import Answer from '@/component/question/answer/index';
-import ColorAnswer from '@/component/question/color-answer';
+import BasicQuestion from '@/component/question/basic-question';
+import ColorQuestion from '@/component/question/color-question';
 import ProgressBar from '@/component/question/progress-bar';
 import { QUESTION_INFO_LIST } from '@/constants/question';
-import type {
-  AnswerHistoryType,
-  AnswerType,
-  CharacterColorType,
-  CharacterType,
-} from '@/types/question';
-import { getColorImageUrl } from '@/utils/answer';
+import type { AnswerHistoryType, AnswerType } from '@/types/question';
 import { setStorage } from '@/utils/storage';
 
 const FIRST_QUESTION_ID = '1';
@@ -28,12 +20,12 @@ function Question() {
   const [answerHistory, setAnswerHistory] = useState<AnswerHistoryType[]>([]);
   const [progressPercent, setProgressPercent] = useState(0);
 
-  const [answerColorStatus, setAnswerColorStatus] = useState('3');
   const [currentQuestionIndex, setCurrentQuestionIndex] =
     useState(FIRST_QUESTION_ID);
 
   const currentQuestion = questions[currentQuestionIndex];
   const isColorQuestion = currentQuestion.type === 'color';
+
   const getNewAnswer = (answer: AnswerType) => {
     const newAnswer: AnswerHistoryType[] = [
       ...answerHistory,
@@ -43,7 +35,6 @@ function Question() {
       },
     ];
     setAnswerHistory(newAnswer);
-    console.log('newAnswer: ', newAnswer);
     return newAnswer;
   };
 
@@ -63,13 +54,6 @@ function Question() {
     setProgressPercent(progressPercent + PROGRESS_STEP);
   };
 
-  const handleAnswerChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const { value } = e.target;
-    setAnswerColorStatus(value);
-  };
-
-  // if (currentQuestion === undefined) return <></>;
-
   return (
     <Wrapper>
       <div>
@@ -81,43 +65,14 @@ function Question() {
         />
       </div>
       <ProgressBar percent={progressPercent} />
-      <div>
-        <ImageBox>
-          {isColorQuestion ? (
-            <Image
-              src={getColorImageUrl(answerColorStatus)}
-              alt="character image"
-              width={250}
-              height={191}
-              placeholder="blur"
-              blurDataURL="/images/blur.webp"
-              priority
-            />
-          ) : (
-            getQuestionCharacterImage(
-              currentQuestion.color,
-              currentQuestion.character,
-            )
-          )}
-        </ImageBox>
-        <GradientBox title={`Q${currentQuestion.id}`}>
-          <QuestionInnerBox>
-            {currentQuestion.content.map((q, index) => (
-              <p key={index + q}>{q}</p>
-            ))}
-          </QuestionInnerBox>
-        </GradientBox>
-      </div>
       {isColorQuestion ? (
-        <ColorAnswer
-          handleAnswerSubmit={handleAnswerClick}
-          answerColorStatus={answerColorStatus}
-          handleAnswerChange={handleAnswerChange}
+        <ColorQuestion
+          currentQuestion={currentQuestion}
+          handleAnswerClick={handleAnswerClick}
         />
       ) : (
-        <Answer
-          type={currentQuestion.type}
-          answers={currentQuestion.answers}
+        <BasicQuestion
+          currentQuestion={currentQuestion}
           handleAnswerClick={handleAnswerClick}
         />
       )}
@@ -132,72 +87,6 @@ const Wrapper = styled.div`
 
   gap: 20px;
 `;
-
-const ImageBox = styled.div`
-  height: 250px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const QuestionInnerBox = styled.div`
-  padding: 20px;
-  font-size: 16px;
-  font-weight: 600;
-  width: 330px;
-
-  color: ${(props) => props.theme.colors.bg};
-`;
-
-const getQuestionCharacterImage = (
-  color: CharacterColorType,
-  character: CharacterType,
-) => {
-  const imageURL = `/problem/${color}_${character}.png`;
-  const { width, height } = getCharacterImageSize(character);
-
-  return (
-    <Image
-      src={imageURL}
-      alt="character image"
-      width={width}
-      height={height}
-      placeholder="blur"
-      blurDataURL={'/images/blur.webp'}
-      priority
-    />
-  );
-};
-
-const getCharacterImageSize = (
-  character: string,
-): {
-  width: number;
-  height: number;
-} => {
-  switch (character) {
-    case 'lumy':
-      return {
-        width: 258,
-        height: 191,
-      };
-    case 'lanny':
-      return {
-        width: 269,
-        height: 191,
-      };
-    case 'cat': //TODO :크기 맞는지 확인
-      return {
-        width: 167,
-        height: 191,
-      };
-    default:
-      return {
-        width: 258,
-        height: 191,
-      };
-  }
-};
 
 export default withLayout(
   Question,
